backend/data/requirements.txt
pdfminer.six==20231228
pandas==2.2.2
python-dotenv==1.0.1
